# REGR: Performance of DataFrame axis=1 reduction ops with EA