@jbrockmendel 2023-03-15
Just realized this is separate from #51923.  Should this take priority over that?  (I think the solution suggested there yesterday is probably the "right" long-term solution)