@jbrockmendel 2023-03-15
I'm generally a believer in "correctness over performance," but this is a cornery enough case im not inclined to argue about it.