@jbrockmendel 2023-03-21
I'm finding myself going through the diff side-by-side with the diff from #51335 and trying to sort out what is and isnt a reversion.  Would it be easier to revert that entirely and then do another PR to un-revert the parts we want to keep?