@datapythonista 2023-04-03
Sorry, my bad, the next release should be 2.0.1, changed it.