@rhshadrach 2023-04-03
> the other pr supersedes this one?

The other PR (#52250) precedes this one, as was created by the request above (https://github.com/pandas-dev/pandas/pull/51955#issuecomment-1478685300)